Making wrong score only negatives

+++ UniTest/src/it/unical/asde2018/unitest/components/controllers/ExamController.java
@@ -72,7 +72,7 @@ public class ExamController {
 
 		Exam e = examService.getExamByInternalID(Integer.parseInt(exam_ID));
 		Question question = new Question(questionTitle, Question_Type.valueOf(questionType),
-				Integer.parseInt(correctScore), Integer.parseInt(wrongScore));
+				Float.parseFloat(correctScore), Float.parseFloat(wrongScore));
 		question.setInternalID(e.getQuestions().size());
 
 		if (Question_Type.valueOf(questionType).equals(Question_Type.SINGLE_CHOICE)) {
